backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.0
pydantic-settings==2.5.2

# Document Processing
pypdf==4.3.1
beautifulsoup4==4.12.3
python-docx==1.1.2
python-multipart==0.0.12
lxml==5.3.0

# OCR
pytesseract==0.3.13
Pillow==10.4.0

# Embeddings & Vector DB
openai==1.51.0
sentence-transformers==3.2.1
qdrant-client==1.12.0

# Multi-Tenancy & Caching
redis==5.2.0
redis[hiredis]==5.2.0

# Observability
prometheus-client==0.21.0
opentelemetry-api==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-instrumentation-fastapi==0.49b1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
pytest-mock==3.14.0
httpx==0.27.2

# Code Quality
black==24.10.0
isort==5.13.2
flake8==7.1.1
pylint==3.3.1
mypy==1.13.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
